requirements.txt
asgiref==3.8.1
boto3==1.35.44
botocore==1.35.44
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-extensions==3.2.3
django-sslserver==0.22
django-storages==1.14.4
filelock==3.16.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
icalendar==6.0.1
idna==3.10
include==0.2.2
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
whitenoise==6.7.0